made the reverse id's use long values even with a performance impact

+++ src/main/java/org/neo4j/batchimport/collections/ReverseRelationshipMap.java
@@ -5,6 +5,6 @@ package org.neo4j.batchimport.collections;
 * @since 27.10.12
 */
 public interface ReverseRelationshipMap {
-     void add(int nodeId, int relId);
-     int[] remove(int nodeId);
+     void add(long nodeId, long relId);
+     long[] remove(long nodeId);
 }
